Extract ETag Reader logic to ReadMethods.Etags.cs with NoInlining
Move GETWITHETAG and GETIFNOTMATCH handling from Reader into a
[NoInlining] HandleEtagReader helper in ReadMethods.Etags.cs.

ETag commands are now delegated at the beginning of Reader before
any other processing, keeping the hot-path Reader method compact.
