+++ libs/server/Storage/Functions/MainStore/ReadMethods.cs
@@ -59,12 +59,12 @@ namespace Garnet.server
             }
 
             var value = srcLogRecord.ValueSpan; // reduce redundant length calculations
-            if (cmd == RespCommand.GETIFNOTMATCH)
-            {
-                if (handleGetIfNotMatch(in srcLogRecord, ref input, ref output, ref readInfo))
-                    return true;
-            }
-            else if (cmd > RespCommandExtensions.LastValidCommand)
+
+            // Delegate ETag commands early to NoInline helpers
+            if (cmd is RespCommand.GETIFNOTMATCH or RespCommand.GETWITHETAG)
+                return HandleEtagReader(in srcLogRecord, ref input, ref output, ref readInfo, cmd, value);
+
+            if (cmd > RespCommandExtensions.LastValidCommand)
             {
                 var valueLength = value.Length;
 
@@ -82,43 +82,12 @@ namespace Garnet.server
                 }
             }
 
-            // Only set ETag state for GETWITHETAG/GETIFNOTMATCH
-            if (cmd is RespCommand.GETWITHETAG or RespCommand.GETIFNOTMATCH)
-            {
-                if (srcLogRecord.Info.HasETag)
-                    ETagState.SetValsForRecordWithEtag(ref functionsState.etagState, in srcLogRecord);
-                CopyRespWithEtagData(value, ref output, srcLogRecord.Info.HasETag, functionsState.memoryPool);
-                ETagState.ResetState(ref functionsState.etagState);
-                return true;
-            }
-
             if (cmd == RespCommand.NONE)
                 CopyRespTo(value, ref output);
             else
                 CopyRespToWithInput(in srcLogRecord, ref input, ref output, readInfo.IsFromPending);
 
             return true;
-        }
-
-        private bool handleGetIfNotMatch<TSourceLogRecord>(in TSourceLogRecord srcLogRecord, ref StringInput input, ref StringOutput dst, ref ReadInfo readInfo)
-            where TSourceLogRecord : ISourceLogRecord
-        {
-            // Any value without an etag is treated the same as a value with an etag
-            long etagToMatchAgainst = input.parseState.GetLong(0);
-
-            long existingEtag = srcLogRecord.ETag;
-
-            if (existingEtag == etagToMatchAgainst)
-            {
-                // write back array of the format [etag, nil]
-                var nilResp = functionsState.nilResp;
-                // *2\r\n: + <numDigitsInEtag> + \r\n + <nilResp.Length>
-                var numDigitsInEtag = NumUtils.CountDigits(existingEtag);
-                WriteValAndEtagToDst(4 + 1 + numDigitsInEtag + 2 + nilResp.Length, nilResp, existingEtag, ref dst, functionsState.memoryPool, writeDirect: true);
-                return true;
-            }
-
-            return false;
         }
 
         /// <summary>
